Create Arduino_Magic_8_Ball
I'm excited to share my latest Arduino project - a Magic 8-Ball with an LCD screen on board! This project includes a 2-row, 16-character LCD display, 20 pre-programmed answers, and uses switch/case statements and the random() function to provide random responses to user questions. With this project, you can ask the Magic 8-Ball anything and receive an answer displayed on the LCD screen. Check out the code and schematics in my repository to create your own version of the Magic 8-Ball with an Arduino!